Update Notices for vue2/vue3 (#481)
Update the Notices file.  The conversion from NPM vue2 to vue3 pulls in a
different set of packages, and their use must be indicated in the Notices.
This also corrects the spelling of the "carbon" package to @carbon as it
notes use of @carbon/motion and @carbon/themes.

Tested: no testing

Signed-off-by: Joseph Reynolds <joseph.reynolds1@ibm.com>